backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
openai>=1.59.0
pypdf==4.3.1
pdfplumber==0.11.4
pydantic==2.9.2
